requirements.txt
flask==3.0.*
flask-cors==4.*
pyjwt==2.*
pillow==10.*
python-dotenv==1.*